gizmo can now move instances around
The gizmo is now able to keep track of any instance positions that have been attached to it. You can click-drag on an axis line to move it along that axis, or press the G key to move it around freely. While in any move mode, the Ctrl key can be held to snap its position to level geometry. Lastly, occurrences of the keyword 'this' were changed in extlib.h to make it C++ compatible.