load and parse actor database from packaged toml
As a test, z64scene now loads and prints the entire actor database at startup. Changes have been made to the Makefile so it packages the toml/ directory with the release (aka its contents are copied into bin/toml/). The executable attempts to determine what directory it is running within so it can find the TOML's. This feature has only been tested on Windows for now.