add EOF marker blobs for trimming last blob
Data blobs are sorted and trimmed, except for the last one in each file. Adding an EOF marker blob will make it so the last blob in the file gets trimmed as well. This will prevent any rare cases where a blob's start + size would attempt to read beyond the end of the file it's referencing.