guarantee most deeply nested DL's are written first
Saved scenes containing multiple mesh headers that point to DL's that point to DL's were having an issue with the pointers to the nested DL's not being updated in all cases. This was because the algorithm relies on the most deeply nested DL's being written first, but they weren't because their parent DL's were being referenced by the room mesh headers and moved back to the front of the list. So now it recursively refreshes each DL each time so those most deeply nested are guaranteed to be at the front of the list in all cases.