tweak blob trimming logic, no cumulative DL size
MM likes to use blank DL's, and it's possible for the blank DL to be the end-DL command of a non-blank DL, causing the end-DL command of a non-blank DL to get trimmed off if this is not accounted for. Also, size of DL's was increasing cumulatively as they were processed multiple times, so that was fixed in this commit as well. MM maps can now be opened and overwritten, then opened and overwritten again without the file layout beginning to deviate. Ran exhaustive tests for both OoT and MM. Valgrind does not report any issues with the tests.