gizmo/gui interop WIP
Actors selected using the drop-down menu are now selected in the viewport, and vice versa. Movements made in the GUI are applied to whichever actor is selected in the viewport. The edited scene can be saved, with new actor positions appearing in-game. I think it would be a little cleaner to refactor this to use the GuiInterop struct instead of the current setter functions. Pushing this WIP as a checkpoint.