new raycast method for selecting instances
Rather than checking for sphere collisions of a fixed radius from an instance's origin, users can now click anywhere on a viewport instance to select it. This method checks for collisions against the individual triangles associated with the instances when clicking. This also allows snapping instances onto other instances when holding Ctrl. When moving an instance, its triangles are ignored during raycasting so it doesn't snap onto itself.