rendercode: SetFaceSnapVector() for specifying angle snap behavior
Automatically rotating an instance based on the angle of the surface it's snapped onto is now configurable per instance using RenderCode. Most instances don't need this, so it's off by default. Also added ClearFaceSnapVector() for turning it back off on instances that do or don't use this feature depending on their properties.