fix corner case actor auto-rotate on-click
After placing an actor (one that does not use auto-rotate) on a surface using Ctrl+V to paste, clicking on a different actor that does have auto-rotate-to-surface functionality would attempt to auto-rotate to match the angle of the surface that the previously-selected actor was pasted onto. This commit fixes that bug.