objects tab: determine which objects are missing/unused
As actors and objects are added/deleted, the 'Objects' tab is updated accordingly to reflect which objects are missing and which objects are unused. The user can select missing objects and click the 'Add' button to add them to the main object list. The user can select unused options and click the 'Delete' button to delete them from the main object list.