scene read/write: optional exit/spawn/header counts
The base game doesn't store the number of elements for exit/spawn/header lists, so the sizes of these blocks must be determined (predicted) using other methods. With this commit, z64scene now writes counts for each, so that they will be known the next time the scene is loaded. This is most important for scenes that are created using z64scene itself. If counts are not specified, it still falls back to trying to predict them.