prevent smaller textures from being coalesced into larger ones
This change prevents smaller textures from being coalesced into larger textures, which was causing the larger textures to be assumed smaller than they actually are, which was causing bytes to be trimmed off the end of them on subsequent save/load cycles. Changing the trim algorithm would be a simpler alternative, if this is ever revisited, but this solution seems more predictable.