99% matching binaries across scene load/save cycles
Big changes to datablobs and scene writing. Tested exhaustively with OoT and MM. Valgrind reports no issues. Of the >200 scenes tested, 3 of them alternate between two checksums after opening/saving/reopening/resaving, and texture/palette data moves around within the file without any data becoming corrupted. The scenes that do this seem to be functional, but I would still prefer to fix this, because consistent and predictable behavior is best. This is a good checkpoint commit.