added flipbook texture animation preview with play/pause button
After clicking on an animated flipbook texture in the 'Segments' tab, the new Play/Pause button can be clicked to preview it. As the animation plays, its individual frames are displayed in the window, and progress is reported as a number between 0.0 and 1.0, where 0.0 and 1.0 are the first and last frames of the animation, respectively.