SceneAnim types: WIP conversion to equivalent types for preview
Converts some SceneAnim types to their AnimatedMaterial equivalents so that different editors for each type are not necessary. Also makes some of them previewable in the viewport. WIP because not all types are converted yet, and they are not yet converted back to appropriate types during export, but this is a good place for a checkpoint commit.