scene write: don't allow coalesced/overlapping blocks
Filesize reductions are achieved by having duplicate data resolve to the same address. However, the size of a block becomes indeterminate when the 'next' block overlaps it, which breaks any logic that is dependent on block size (for example: tables without a defined length). The scene writer still takes advantage of duplicate blocks in order to save space, but now blocks are only matched to other blocks of the same type and size, and never matched to subsections within other blocks. This matching method is also a little bit faster, though that wasn't the primary motivation.