preserve unreferenced cameras across scene load/save
Camera list length is not defined by scenes, and so must be determined using other methods. Checking camera indices referenced by collision surface types works for most cases, but there are still rare instances of cameras not referenced in this way. This new method carefully inspects the data at the end of the camera list to see is there more data resembling camera data. Any data erroneously added to the camera list that is part of the subsequent block gets trimmed away using a post-sort callback, so this should be very safe and reliable.