path point list, can extend paths, reverse/rotate
Individual path points are now listed in and able to be selected from the UI. In the viewport, the 'V' key can be pressed to extend a path from either direction (beginning or end). Also added helpers for reversing and rotating arrays, which make it easy to reverse paths, as well as change which point on the path is intended to be the starting point.