path editor: support closed aka looping paths
When a path has more than 3 points, UI appears for toggling whether it is a 'closed' path. A closed path has its first and last points welded together so that it loops seamlessly. Unlike open paths, changing which point of a looping path is its start point does not affect its shape.