fix crash on OOB animation data
Animation data in OoT/MM does not define a fixed number of bones, so there is no reliable way to tell which animations are intended to be used by which skeletons. For this reason, it is possible to load the wrong animation with the wrong skeleton, and for OOB reads to occur. The OOB checks I added in this commit prevent that.