feature: move scene textures to external object
Asks user where to save the object. A z64yartool recipe will be created in the same file as the generated object. When given the recipe, z64yartool can dump textures from that object for editing, as well as inject them back in. Requires z64yartool v1.1.0 or above. Additionally, the scene and room files will have their textures removed from them, and all references to those textures are updated to point to the external object. Intended to be used with this feature in z64rom 1.7.0 and above: https://github.com/z64utils/z64rom/pull/3